+++ packages/webpack/build.config.ts
@@ -3,6 +3,7 @@ import { defineBuildConfig } from 'unbuild'
 export default defineBuildConfig({
   entries: [
     'src/index',
+    'src/rspack',
   ],
   clean: true,
   declaration: true,
