+++ packages/webpack/src/rspack.ts
@@ -1,0 +1,10 @@
+import type { UserConfigDefaults } from '@unocss/core'
+import type { WebpackPluginOptions } from '.'
+import { unplugin } from './unplugin'
+
+export function UnoCSSRspackPlugin<Theme extends object>(
+  configOrPath?: WebpackPluginOptions<Theme> | string,
+  defaults?: UserConfigDefaults,
+) {
+  return unplugin(configOrPath, defaults).rspack()
+}
